backend/requirements.txt
Django==3.2.19
django-colorful==1.3
pillow==9.5.0
djangorestframework==3.14.0
django-filter==23.2
djoser==2.1.0
django-cors-headers==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
